extension: allow user to specify an editor through an option page
This patch adds an options/settings page to the extension, allowing the
user to pass the editor executable path and args or directly the shell
command used to run that editor (including the args).

The editor set in the option page is passed from the extension to the
native app through the same custom header message we used to handle
paged responses, but using the field "Command:". The command is
preserved upon addon reload, since it's stored in Thunderbird's local
storage.

Alongside this changes, I used the opportunity to add the icon for the
extension: a beatiful and simple bird + paper.