app: rework payload logic for paged messages
The payload length checked by the WebExtension protocol considers the
size of the encoded (json) message, and because of that plus the fact we
support paged messages, multiples back'n'forth were required to get the
right size of the encoded payload each time we split the plaintext
payload. With that the code was getting bigger and confusing.

This patch fixes it by drastically decreasing the length limit for the
plain text message, which allow us to basically ignore the encoded
payload length and only consider the plaintext, easing the paged message
handling. At the same time, both encoded payload and number of pages are
now set at setPayload() time, instead of on sendMessage().