extension: fix check for port error on connect
The port connection is an async call that doesn't return a promise, thus
we can't wait on them nor directly check for a port error right after
the connectNative() call, but we can add a listener to the onDisconnect
event, which is triggered in case the connection failed for whatever
reason. This patch adds the error checker as a handler to the
onDisconnect() event and also make sure to check both runtime.lastError
and port.error to make sure we catch the error (the later isn't
supported on chrome based extension).