extension: use plaintext even though handling HTML
tbed still doesn't support HTML text, however, we still can grab the
plaintext version of it for opening and editing on the external editor.
The drawback is that the text returned by the editor will be plaintext
with no HTML at all, completely changing the previous email content
(considering the case where the user is replying to an HTML body). To
help a little, this commit converts the plaintext from the response to a
simple HTML using the internal JavaScript engine by creating a random
HTML element and setting its innerHTML to the plaintext and reading the
outerHTML from that, which contains that text with additional HTML tags
and elements to better represent the original response plaintext.

For now, it seems acceptable, we just need to make sure the user is
aware of such behavior, possibly by prompting a warning or something
like that (still TBD).